==> 1_create_pitchers_and_batters.ipynb <==
import pandas as pd
import numpy as np

lhh_vs_lhp = pd.read_csv('./Datasets/Downloads/Splits Leaderboard Data vs LHP as LHH.csv')
lhh_vs_rhp = pd.read_csv('./Datasets/Downloads/Splits Leaderboard Data vs RHP as LHH.csv')
rhh_vs_rhp = pd.read_csv('./Datasets/Downloads/Splits Leaderboard Data as RHH vs RHP.csv')
rhh_vs_lhp = pd.read_csv('./Datasets/Downloads/Splits Leaderboard Data as RHH vs LHP.csv')

lhh_vs_lhp['bats'] = 'L'
lhh_vs_lhp['pitch'] = 'L'
lhh_vs_rhp['bats'] = 'L'
lhh_vs_rhp['pitch'] = 'R'
rhh_vs_rhp['bats'] = 'R'
rhh_vs_rhp['pitch'] = 'R'
rhh_vs_lhp['bats'] = 'R'
rhh_vs_lhp['pitch'] = 'L'

batters = pd.concat([lhh_vs_lhp, lhh_vs_rhp, rhh_vs_lhp, rhh_vs_rhp])

lhp_vs_lhh = pd.read_csv('./Datasets/Splits Leaderboard Data vs LHH as LHP.csv')
lhp_vs_rhh = pd.read_csv('./Datasets/Splits Leaderboard Data vs RHH as LHP.csv')
rhp_vs_lhh = pd.read_csv('./Datasets/Splits Leaderboard Data vs LHH as RHP.csv')
rhp_vs_rhh = pd.read_csv('./Datasets/Splits Leaderboard Data vs RHH as RHP.csv')

lhp_vs_lhh['throws'] = 'L'
lhp_vs_lhh['batter'] = 'L'
lhp_vs_rhh['throws'] = 'L'
lhp_vs_rhh['batter'] = 'R'
rhp_vs_lhh['throws'] = 'R'
rhp_vs_lhh['batter'] = 'L'
rhp_vs_rhh['throws'] = 'R'
rhp_vs_rhh['batter'] = 'R'

pitchers = pd.concat([lhp_vs_lhh, lhp_vs_rhh, rhp_vs_lhh, rhp_vs_rhh])

batters.to_csv('./Datasets/batter_stats.csv', index = False)
pitchers.to_csv('./Datasets/pitcher_stats.csv', index = False)

lineups = pd.read_csv('./Datasets/Downloads/Lineups_2021_05_29.csv')

team_list_5_29 = []
for team in lineups['team code']:
    if team not in team_list_5_29:
        team_list_5_29.append(team)
matchups_5_29 = {}
for i in range(0, len(team_list), 2):
    matchups_5_29[team_list_5_29[i]] = team_list_5_29[i+1]
    matchups_5_29[team_list_5_29[i+1]] = team_list_5_29[i]

matchups_5_29

def inning(players, pitcher):
    outs = 0
    bases = [0,0,0]
