==> 3_create_pitcher_train_data.ipynb <==
import pandas as pd
import numpy as np
import os as os
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.ensemble import RandomForestRegressor
from sklearn.linear_model import LinearRegression, LassoCV, Lasso
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
from sklearn.model_selection import train_test_split
from sklearn.decomposition import PCA
from sklearn.metrics import mean_squared_error

pitcher_csv = [file for file in os.listdir('./merged_data') if 'pitchers' in file]

li = []
for csv in pitcher_csv:
    li.append(pd.read_csv('./merged_data/' + csv))
pitchers = pd.concat(li, axis = 0, ignore_index = True)

# CSV of opposing teams average runs scored
runs = pd.read_csv('./Datasets/team_runs.csv')

# Formatting the Oppt column to merge opponents average runs
pitchers['Oppt'] = [opt.split()[-1].upper() for opt in pitchers['Oppt']]

# Merging in the opposing teams average runs(Later found to not be significant)
pitch_df = pd.merge(left = pitchers, right = runs, left_on = 'Oppt',
        right_on = 'team').drop(columns = ['team', 'Unnamed: 0', 'Tm Runs', 'Opp Runs'])

pitch_df.dropna(subset = ['DK pts','DK sal'])

pitch_df.columns

# Selecting features to be exported to google colab
# The features will be trimmed down further in google colab
p_feat = ['num_L_bats', 'num_R_bats',
       'is_nl', 'k_9_l', 'bb_9_l', 'k_bb_l', 'hr_9_l', 'k_pct_l', 'bb_pct_l',
       'k_bb_pct_l', 'bat_avg_l', 'whip_l', 'babip_l', 'lob_pct_l', 'fip_l',
       'xfip_l', 'k_9_r', 'bb_9_r', 'k_bb_r', 'hr_9_r', 'k_pct_r', 'bb_pct_r',
       'k_bb_pct_r', 'bat_avg_r', 'whip_r', 'babip_r', 'lob_pct_r', 'fip_r',
       'xfip_r', 'k_9_l_tot', 'bb_9_l_tot', 'k_bb_l_tot', 'hr_9_l_tot',
       'k_pct_l_tot', 'bb_pct_l_tot', 'k_bb_pct_l_tot', 'bat_avg_l_tot',
       'whip_l_tot', 'babip_l_tot', 'lob_pct_l_tot', 'fip_l_tot', 'xfip_l_tot',
       'k_9_r_tot', 'bb_9_r_tot', 'k_bb_r_tot', 'hr_9_r_tot', 'k_pct_r_tot',
       'bb_pct_r_tot', 'k_bb_pct_r_tot', 'bat_avg_r_tot', 'whip_r_tot',
       'babip_r_tot', 'lob_pct_r_tot', 'fip_r_tot', 'xfip_r_tot', 'DK sal']
X = pitch_df[p_feat]
y = pitch_df['DK pts']

# preparing to export data to use in neural network in google colab
p_feat.append('DK pts')
final_pitch = pitch_df[p_feat]
final_pitch.head()

final_pitch.to_csv('./merged_data/final_pitch.csv', index = False)
